Fixes #10052 - Enabling consumer reg for atomic hosts
Main intent of this commit is to setup rhsm conf on an atomic host to
setup consumer registration (a role similar to the one played by the
consumer bootstrap rpm).

Artifacts produced

1) /var/www/html/pub/katello-rhsm-consumer on the satellite. This script
contains the code needed to setup rhsm conf on both atomic as well as
non atomic machines. The contents of this script came out of the template
used for post install script in the bootstrap consumer rpm. It has been
tuned to work for both atomic/non-atomic machines

2) Update to the bootstrap rpm to copy the new script to /usr/bin and
run that as a part of post install.